+++ docs/pages/_en/1.0/reference/release-notes.md
@@ -14,6 +14,7 @@ group: Deployment & Configs
 
 - Added 'Component Family Mesh' component.
 - Updated 'View Range' to be also a setter.
+- Now 'Graphical Element' is able to grab selected linked elements when in 'Active Selection'.
 
 {% endcapture %}
 {% include ltr/release_header_next.html title="Upcoming Changes" note=rc_release_notes %}